Moved taskbar, misc styling
